make openUrl function generic
