+++ scripts/delayed.js
@@ -12,7 +12,7 @@ import {
   pushProductsToDataLayer,
   pushToDataLayer,
   getEnvironment,
-  getOperatingSystem,
+  openUrlForOs,
 } from './scripts.js';
 import { loadBreadcrumbs } from './breadcrumbs.js';
 
@@ -41,41 +41,11 @@ pushToDataLayer('page loaded');
 loadBreadcrumbs();
 
 // Get the open URL for the user's OS
-const openUrlMacos = getMetadata('open-url-macos');
-const openUrlWindows = getMetadata('open-url-windows');
-const openUrlAndroid = getMetadata('open-url-android');
-const openUrlIos = getMetadata('open-url-ios');
+const urlMacos = getMetadata('open-url-macos');
+const urlWindows = getMetadata('open-url-windows');
+const urlAndroid = getMetadata('open-url-android');
+const urlIos = getMetadata('open-url-ios');
 
-if (openUrlMacos || openUrlWindows || openUrlAndroid || openUrlIos) {
-  // Get user's operating system
-  const { userAgent } = navigator;
-  const userOS = getOperatingSystem(userAgent);
-
-  // Open the appropriate URL based on the OS
-  let openUrl;
-  switch (userOS) {
-    case 'MacOS':
-      openUrl = openUrlMacos;
-      break;
-    case 'Windows 10':
-    case 'Windows 8':
-    case 'Windows 7':
-    case 'Windows Vista':
-    case 'Windows XP':
-    case 'Windows 2000':
-      openUrl = openUrlWindows;
-      break;
-    case 'Android':
-      openUrl = openUrlAndroid;
-      break;
-    case 'iOS':
-      openUrl = openUrlIos;
-      break;
-    default:
-      openUrl = null; // Fallback or 'Unknown' case
-  }
-
-  if (openUrl) {
-    window.open(openUrl, '_self');
-  }
+if (urlMacos || urlWindows || urlAndroid || urlIos) {
+  openUrlForOs(urlMacos, urlWindows, urlAndroid, urlIos);
 }
